requirements.txt
streamlit>=1.22.0,<2.0.0
openai>=0.27.6,<1.0.0
sentence-transformers>=2.3.0,<3.0.0
numpy>=1.23.5,<2.0.0
python-dotenv>=1.0.0,<2.0.0
torch>=1.13.1,<2.0.0